Time: 0 ms (100%), Space: 8 MB (2.16%) - LeetHub

+++ 1922-count-good-numbers/1922-count-good-numbers.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long MOD=1e9+7;
-    long long power(long long x,long long n){
+    long long power(int x,long long n){
         if(n==0){
             return 1;
         }
